refactor: move CreateSubscriptionRegistrationLink to its own file
The tool belongs to a separate resource (subscription_registration),
not payment_links. Extract it into:
- pkg/razorpay/subscription_registration_links.go
- pkg/razorpay/subscription_registration_links_test.go

Register it under a new `subscription_registration_links` toolset in
tools.go instead of the payment_links toolset.
