+++ pkg/razorpay/tools.go
@@ -42,9 +42,15 @@ func NewToolSets(
 		AddWriteTools(
 			CreatePaymentLink(obs, client),
 			CreateUpiPaymentLink(obs, client),
-			CreateSubscriptionRegistrationLink(obs, client),
 			ResendPaymentLinkNotification(obs, client),
 			UpdatePaymentLink(obs, client),
+		)
+
+	subscriptionRegistrationLinks := toolsets.NewToolset(
+		"subscription_registration_links",
+		"Razorpay Subscription Registration Links related tools").
+		AddWriteTools(
+			CreateSubscriptionRegistrationLink(obs, client),
 		)
 
 	orders := toolsets.NewToolset("orders", "Razorpay Orders related tools").
@@ -119,6 +125,7 @@ func NewToolSets(
 	toolsetGroup.AddToolset(checkoutIntegration)
 	toolsetGroup.AddToolset(payments)
 	toolsetGroup.AddToolset(paymentLinks)
+	toolsetGroup.AddToolset(subscriptionRegistrationLinks)
 	toolsetGroup.AddToolset(orders)
 	toolsetGroup.AddToolset(refunds)
 	toolsetGroup.AddToolset(payouts)
